src/securia_api/requirements.txt
envyaml
requests
pyOpenSSL
json2html
fastapi
cryptography
uvicorn
prometheus_client
schedule
# kubernetes
confluent-kafka
# kafka-python
kafka-python-ng
prometheus-client
sqlalchemy
pydantic
psycopg2